Start working on database, simplify fetching data with moshi adapter

+++ app/src/main/java/com/currenciesapp/model/CurrencyItem.kt
@@ -3,7 +3,7 @@ package com.currenciesapp.model
 data class CurrencyItem(
     val code: String,
     val fullName: String,
-    val rate: Float
+    val rate: Double
 )
 
 fun Currency.toItem() =
